requirements.txt
python-dotenv
torch
transformers
langchain-google-genai
langchain-core
speechrecognition
pyttsx3
pyqt5
pyaudio